+++ nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=/home/ducnam/snap/netbeans/common/data/11.2/build.properties
+user.properties.file=C:\\Users\\LEGION\\AppData\\Roaming\\NetBeans\\11.1\\build.properties
