+++ nbproject/private/private.xml
@@ -3,18 +3,16 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/home/ducnam/Btlnew/src/Model/MonHoc.java</file>
-            <file>file:/home/ducnam/Btlnew/src/View/ImageIcon/user.png</file>
-            <file>file:/home/ducnam/Btlnew/src/View/EditProflie.java</file>
-            <file>file:/home/ducnam/Btlnew/src/View/RegisterFrame.java</file>
-            <file>file:/home/ducnam/Btlnew/src/View/QuanLi/QuanLiFrame.java</file>
-            <file>file:/home/ducnam/Btlnew/src/View/GiangVien/GVFrame.java</file>
-            <file>file:/home/ducnam/Btlnew/src/Controller/LoginController.java</file>
-            <file>file:/home/ducnam/Btlnew/src/Model/GiangVien.java</file>
-            <file>file:/home/ducnam/Btlnew/src/View/MainFrame.java</file>
-            <file>file:/home/ducnam/Btlnew/src/View/LoginFrame.java</file>
-            <file>file:/home/ducnam/Btlnew/src/View/SinhVien/SVFrame.java</file>
-            <file>file:/home/ducnam/Btlnew/src/Model/SinhVien.java</file>
+            <file>file:/C:/Users/LEGION/Documents/NetBeansProjects/BaiTapLon/src/DAO/DBQuery.java</file>
+            <file>file:/C:/Users/LEGION/Documents/NetBeansProjects/BaiTapLon/src/Model/GiangVien.java</file>
+            <file>file:/C:/Users/LEGION/Documents/NetBeansProjects/BaiTapLon/src/View/QuanLi/SvDisplayPanel.java</file>
+            <file>file:/C:/Users/LEGION/Documents/NetBeansProjects/BaiTapLon/src/View/QuanLi/SinhVien/GiangVienComponent.java</file>
+            <file>file:/C:/Users/LEGION/Documents/NetBeansProjects/BaiTapLon/src/Model/SinhVien.java</file>
+            <file>file:/C:/Users/LEGION/Documents/NetBeansProjects/BaiTapLon/src/View/QuanLi/GvDisplayPanel.java</file>
+            <file>file:/C:/Users/LEGION/Documents/NetBeansProjects/BaiTapLon/src/DAO/ProfessorQuery.java</file>
+            <file>file:/C:/Users/LEGION/Documents/NetBeansProjects/BaiTapLon/src/Controller/QuanLi/QLViewController.java</file>
+            <file>file:/C:/Users/LEGION/Documents/NetBeansProjects/BaiTapLon/src/DAO/StudentQuery.java</file>
+            <file>file:/C:/Users/LEGION/Documents/NetBeansProjects/BaiTapLon/src/View/QuanLi/QuanLiFrame.java</file>
         </group>
     </open-files>
 </project-private>
