+++ src/Model/MonHoc.java
@@ -17,17 +17,20 @@ public class MonHoc implements Serializable{
     private int maMon;
     private String tenMon;
     private int sotinchi;
-    private int maGv;
-    private String phong;
-    private String thoigian;
-    private String tietbd;
-    private String ngay;
+    private String khoa;
 
-    public MonHoc(int maMon, String tenMon, int sotinchi, int maGv) {
+    public String getKhoa() {
+        return khoa;
+    }
+
+    public void setKhoa(String khoa) {
+        this.khoa = khoa;
+    }
+    public MonHoc(int maMon, String tenMon, int sotinchi,String khoa) {
         this.maMon = maMon;
         this.tenMon = tenMon;
         this.sotinchi = sotinchi;
-        this.maGv = maGv;
+        this.khoa = khoa;
     }
 
     public MonHoc() {
@@ -57,53 +60,8 @@ public class MonHoc implements Serializable{
         this.sotinchi = sotinchi;
     }
 
-    public int getMaGv() {
-        return maGv;
-    }
 
-    public void setMaGv(int maGv) {
-        this.maGv = maGv;
-    }
-        public String getPhong() {
-        return phong;
-    }
+    
 
-    public void setPhong(String phong) {
-        this.phong = phong;
-    }
 
-    public String getThoigian() {
-        return thoigian;
-    }
-
-    public void setThoigian(String thoigian) {
-        this.thoigian = thoigian;
-    }
-
-    public String getTietbd() {
-        return tietbd;
-    }
-
-    public void setTietbd(String tietbd) {
-        this.tietbd = tietbd;
-    }
-
-    public String getNgay() {
-        return ngay;
-    }
-
-    public void setNgay(String ngay) {
-        this.ngay = ngay;
-    }
-
-    public MonHoc(int maMon, String tenMon, int sotinchi, int maGv, String phong, String thoigian, String tietbd, String ngay) {
-        this.maMon = maMon;
-        this.tenMon = tenMon;
-        this.sotinchi = sotinchi;
-        this.maGv = maGv;
-        this.phong = phong;
-        this.thoigian = thoigian;
-        this.tietbd = tietbd;
-        this.ngay = ngay;
-    }
 }
